Move Numba scalar tests to test_scalar
